Converts state to an Object
Before, the state was just a string. Making it an Object means
I can add other properties, specifically: the values of the bit
fields. I do wonder if I can just set ONE of the properties, or
does the "setter" need to set all the properties.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object

NOTE: I'm following:

https://www.twilio.com/blog/react-choose-functional-components

and:

https://reactjs.org/docs/state-and-lifecycle.html