Makes binary number appear in checkbox field
This was the goal, and I achieved it! I can probably remove
the Binary string, but it's helpful to have.

There should be some validation if the input value is above
255 or below 0. That could be next. The other thing that is
next is to make a handler for the checkboxes so that updating
that will also update the input value.

React state for the win!