Makes checkboxes work
After some time cogitating about it, I found out
that I'm just writing another function to adjust
the state. The only hard part was working the
calculation of the new number, based on which checkbox
was toggled. The order of the bits does NOT correspond
with the order of the displayed checkboxes, so I have
to calculate the "actual_bit". That is:

A = checkbox index selected
B = actual bit this represents

A B
0 7
1 6
2 5
3 4
4 3
5 2
6 1
7 0